added an unfllow command to delete feed follows

+++ internal/database/feed_follows.sql.go
@@ -73,11 +73,16 @@ func (q *Queries) CreateFeedFollow(ctx context.Context, arg CreateFeedFollowPara
 }
 
 const deleteFeedFollow = `-- name: DeleteFeedFollow :exec
-delete from feed_follows
+delete from feed_follows where user_id = $1 and feed_id = $2
 `
 
-func (q *Queries) DeleteFeedFollow(ctx context.Context) error {
-	_, err := q.db.ExecContext(ctx, deleteFeedFollow)
+type DeleteFeedFollowParams struct {
+	UserID uuid.UUID
+	FeedID uuid.UUID
+}
+
+func (q *Queries) DeleteFeedFollow(ctx context.Context, arg DeleteFeedFollowParams) error {
+	_, err := q.db.ExecContext(ctx, deleteFeedFollow, arg.UserID, arg.FeedID)
 	return err
 }
